move tracing-processor files.
